feat-#16: move repositories to domain
