+++ src/item/application/commands/handlers/create-item.handler.ts
@@ -3,7 +3,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 
 import { ErrorValidationService, RpcExceptionService } from '../../../../utils';
 import { ItemEntity } from '../../../infrastructure/entities';
-import { ItemRepository } from '../../../repositories';
+import { ItemRepository } from '../../../domain/repositories';
 import { ItemWriteRepository } from '../../../infrastructure/repositories';
 import { CreateItemCommand } from '../impl';
