+++ src/item/domain/domain.module.ts
@@ -1,4 +1,13 @@
-import { Module } from '@nestjs/common';
+import { Global, Module } from '@nestjs/common';
+import { ItemRepository } from './repositories';
 
-@Module({})
+@Global()
+@Module({
+    providers: [
+        ItemRepository,
+    ],
+    exports: [
+        ItemRepository,
+    ]
+})
 export class DomainModule {}
